lin/requirements.txt
# Core dependencies
python-can>=4.0.0
typing_extensions>=3.10.0

# Development dependencies (install with pip install -r requirements-dev.txt)
# pytest>=6.0
# pytest-timeout
# pytest-cov
# black
# mypy
# ruff